requirements.txt
aiosqlite==0.20.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.3
dateutils==0.6.12
dnspython==2.6.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fake-useragent==1.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
loguru==0.7.2
openpyxl==3.1.4
outcome==1.3.0.post0
packaging==24.1
prometheus_client==0.20.0
pyaes==1.6.1
pyasn1==0.6.0
pycountry==22.1.10
pymongo==4.6.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
rsa==4.9
selenium==4.16.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
Telethon==1.34.0
trio==0.25.1
trio-websocket==0.11.1
tweety-ns==1.1.7
twscrape==0.13
typing_extensions==4.12.2
urllib3==2.2.2
webdriver-manager==4.0.1
wsproto==1.2.0
